Move Lua unit wrapper to own file
