+++ src/scripting/application_lua_kernel.cpp
@@ -30,7 +30,6 @@
 #include "config.hpp"
 #include "game_errors.hpp"
 #include "log.hpp"
-#include "scripting/lua_api.hpp"
 #include "scripting/lua_common.hpp"
 #include "scripting/lua_cpp_function.hpp"
 #include "scripting/lua_fileops.hpp"
