+++ src/scripting/lua_gui2.cpp
@@ -41,7 +41,6 @@
 
 #include "config.hpp"
 #include "log.hpp"
-#include "scripting/lua_api.hpp"        // for luaW_toboolean, etc
 #include "scripting/lua_common.hpp"
 #include "scripting/lua_types.hpp"      // for getunitKey, dlgclbkKey, etc
 #include "serialization/string_utils.hpp"
