+++ src/scripting/lua_pathfind_cost_calculator.hpp
@@ -1,7 +1,6 @@
 #pragma once
 #include "lua/lua.h"
 #include "map/location.hpp"
-#include "scripting/lua_api.hpp"
 #include "pathfind/pathfind.hpp"
 
 /**
