+++ src/scripting/lua_unit.cpp
@@ -12,32 +12,16 @@
    See the COPYING file for more details.
 */
 
-#include "lua_api.hpp"
+#include "lua_unit.hpp"
 #include "lua_types.hpp"
 
-#include "lua_jailbreak_exception.hpp"  // for tlua_jailbreak_exception
-
-#include "chat_events.hpp"              // for chat_handler, etc
-#include "config.hpp"
-#include "display_chat_manager.hpp"
 #include "game_board.hpp"
-#include "game_display.hpp"
 #include "log.hpp"
 #include "map/location.hpp"             // for map_location
 #include "resources.hpp"
 #include "scripting/lua_common.hpp"
-#include "tstring.hpp"
 #include "units/unit.hpp"
 #include "units/map.hpp"
-#include "variable.hpp"
-
-#include <boost/variant/static_visitor.hpp>
-#include <map>                          // for map<>::key_type
-#include <new>                          // for operator new
-#include <ostream>                      // for operator<<, basic_ostream, etc
-#include <utility>                      // for pair
-
-#include <string>
 
 #include "lua/lauxlib.h"
 #include "lua/lua.h"                    // for lua_State, lua_settop, etc
@@ -45,92 +29,6 @@
 static lg::log_domain log_scripting_lua("scripting/lua");
 #define LOG_LUA LOG_STREAM(info, log_scripting_lua)
 #define ERR_LUA LOG_STREAM(err, log_scripting_lua)
-
-void chat_message(std::string const &caption, std::string const &msg)
-{
-	if (!resources::screen) return;
-	resources::screen->get_chat_manager().add_chat_message(time(nullptr), caption, 0, msg,
-		events::chat_handler::MESSAGE_PUBLIC, false);
-}
-
-// To silence "no prototype" warnings
-void push_error_handler(lua_State *L);
-int luaW_pcall_internal(lua_State *L, int nArgs, int nRets);
-
-void push_error_handler(lua_State *L)
-{
-	lua_pushlightuserdata(L, executeKey);
-	lua_getglobal(L, "debug");
-	lua_getfield(L, -1, "traceback");
-	lua_remove(L, -2);
-	lua_rawset(L, LUA_REGISTRYINDEX);
-	lua_pop(L, 1);
-}
-
-int luaW_pcall_internal(lua_State *L, int nArgs, int nRets)
-{
-	// Load the error handler before the function and its arguments.
-	lua_pushlightuserdata(L
-						  , executeKey);
-
-	lua_rawget(L, LUA_REGISTRYINDEX);
-	lua_insert(L, -2 - nArgs);
-
-	int error_handler_index = lua_gettop(L) - nArgs - 1;
-
-	// Call the function.
-	int errcode = lua_pcall(L, nArgs, nRets, -2 - nArgs);
-
-	// Remove the error handler.
-	lua_remove(L, error_handler_index);
-
-	tlua_jailbreak_exception::rethrow();
-
-	return errcode;
-}
-
-#ifdef _MSC_VER
-#pragma warning (push)
-#pragma warning (disable: 4706)
-#endif
-bool luaW_pcall(lua_State *L, int nArgs, int nRets, bool allow_wml_error)
-{
-	int res = luaW_pcall_internal(L, nArgs, nRets);
-
-	if (res)
-	{
-		/*
-		 * When an exception is thrown which doesn't derive from
-		 * std::exception m will be nullptr pointer.
-		 */
-		char const *m = lua_tostring(L, -1);
-		if(m) {
-			if (allow_wml_error && strncmp(m, "~wml:", 5) == 0) {
-				m += 5;
-				char const *e = strstr(m, "stack traceback");
-				lg::wml_error() << std::string(m, e ? e - m : strlen(m));
-			} else if (allow_wml_error && strncmp(m, "~lua:", 5) == 0) {
-				m += 5;
-				char const *e = nullptr, *em = m;
-				while (em[0] && ((em = strstr(em + 1, "stack traceback"))))
-#ifdef _MSC_VER
-#pragma warning (pop)
-#endif
-					e = em;
-				chat_message("Lua error", std::string(m, e ? e - m : strlen(m)));
-			} else {
-				ERR_LUA << m << '\n';
-				chat_message("Lua error", m);
-			}
-		} else {
-			chat_message("Lua caught unknown exception", "");
-		}
-		lua_pop(L, 2);
-		return false;
-	}
-
-	return true;
-}
 
 lua_unit::~lua_unit()
 {
@@ -224,6 +122,7 @@ unit_ptr luaW_checkunit_ptr(lua_State *L, int index, bool only_on_map)
 	if (!u) luaL_typerror(L, index, "unit");
 	return u;
 }
+
 unit& luaW_checkunit(lua_State *L, int index, bool only_on_map)
 {
 	unit* u = luaW_tounit(L, index, only_on_map);
