+++ src/scripting/mapgen_lua_kernel.cpp
@@ -17,7 +17,6 @@
 #include "config.hpp"
 #include "game_errors.hpp"
 #include "log.hpp"
-#include "scripting/lua_api.hpp"
 #include "scripting/lua_common.hpp"
 #include "scripting/lua_rng.hpp"
 #include "scripting/lua_pathfind_cost_calculator.hpp"
